pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oncopulse"
version = "0.1.0"
description = "Oncology Research Inbox"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "streamlit>=1.35.0",
  "requests>=2.31.0",
  "PyYAML>=6.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
]

[tool.setuptools.packages.find]
include = ["oncopulse*"]
